+++ db/migrate/20140211183030_add_additional_admin_only_fields_to_applicants.rb
@@ -1,0 +1,7 @@
+class AddAdditionalAdminOnlyFieldsToApplicants < ActiveRecord::Migration
+  def change
+    add_column :applicants, :curriculum_advisor, :string
+    add_column :applicants, :most_recent_curriculum_advisor_meeting_date, :date
+    add_column :applicants, :past_curriculum_advisor_meetings, :text
+  end
+end